requirements.txt
streamlit>=1.28.0
folium>=0.14.0
pillow>=10.0.0
exifread>=3.0.0
opencv-python>=4.8.0

geopy>=2.3.0
